requirements.txt
asgiref==3.8.1
Django==5.1.6
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.11.0
Markdown==3.9
pillow==11.3.0
psycopg2==2.9.11
pygame==2.6.1
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
